aaboot

+++ avr/AAMini/AAmini.h
@@ -87,8 +87,8 @@ volatile uint16_t io_timer;			// 1 ms timer counter
 
 // Prototypes
 void SendInformationEvent( uint8_t idx, uint8_t eventClass, uint8_t eventTypeId );;
-int readEEPROM( uint8_t addr );
-int writeEEPROM( uint8_t addr, uint8_t data );
+int readEEPROM( uint16_t addr );
+int writeEEPROM( uint16_t addr, uint8_t data );
 
 void SendInformationEventExtended
 	(uint8_t priority, uint8_t zone, uint8_t subzone, uint8_t idx,
